requirements.txt
pytest
selenium
webdriver-manager
pytest-html
pytest-xdist
allure-pytest
openpyxl
mysql-connector-python
pypdf
pdfplumber